+++ app/Http/Controllers/Statuses/StatusController.php
@@ -2,11 +2,11 @@
 namespace EQM\Http\Controllers\Statuses;
 
 use EQM\Core\Files\Uploader;
+use EQM\Http\Controllers\Controller;
 use EQM\Models\Statuses\Requests\PostStatus;
 use EQM\Models\Statuses\Status;
 use EQM\Models\Statuses\StatusCreator;
 use EQM\Models\Statuses\StatusRepository;
-use Illuminate\Routing\Controller;
 
 class StatusController extends Controller
 {
@@ -59,6 +59,8 @@ class StatusController extends Controller
      */
     public function edit(Status $status)
     {
+        $this->authorize('edit-status', $status);
+
         return view('statuses.edit', compact('status'));
     }
 
@@ -69,6 +71,8 @@ class StatusController extends Controller
      */
     public function update(PostStatus $request, Status $status)
     {
+        $this->authorize('edit-status', $status);
+
         $this->statuses->update($status, $request->all());
 
         return redirect()->route('home');
@@ -80,6 +84,8 @@ class StatusController extends Controller
      */
     public function delete(Status $status)
     {
+        $this->authorize('delete-status', $status);
+
         $status->delete();
 
         return redirect()->back();
